+++ neurst/data/datasets/__init__.py
@@ -15,3 +15,4 @@ for file in os.listdir(models_dir):
         module = importlib.import_module('neurst.data.datasets.' + model_name)
 
 importlib.import_module("neurst.data.datasets.audio")
+importlib.import_module("bytedseq.data.datasets.data_sampler")