+++ neurst/tasks/seq2seq.py
@@ -271,7 +271,12 @@ class Seq2Seq(Task):
         return [SequenceTokenMetricLayer("src"), SequenceTokenMetricLayer("trg"),
                 BatchCountMetricLayer("src")]
 
-    def get_eval_metric(self, args, name="metric"):
-        """ Returns a neurst.metrics.metric.Metric object for evaluation."""
+    def get_eval_metric(self, args, name="metric", ds=None):
+        """ Returns a bytedseq.metrics.metric.Metric object for evaluation."""
+        if ds is None:
+            return None
+        elif hasattr(ds, "trg_lang") and ds.trg_lang is not None:
+            return build_metric(args[name + ".class"], language=ds.trg_lang,
+                                **args[name + ".params"])
         return build_metric(args[name + ".class"], language=self._trg_data_pipeline.meta["language"],
                             **args[name + ".params"])