+++ neurst/tasks/speech2text.py
@@ -16,12 +16,13 @@ from typing import Tuple
 
 import tensorflow as tf
 from absl import logging
+from bytedseq.layers.metric_layers.token_metric_layers import (AudioFramesMetricLayer, BatchCountMetricLayer,
+                                                               SequenceTokenMetricLayer)
 
 from neurst.data import dataset_utils
 from neurst.data.data_pipelines import DataPipeline, build_data_pipeline
 from neurst.data.data_pipelines.transcript_data_pipeline import TranscriptDataPipeline
 from neurst.data.datasets import Dataset
-from neurst.layers.metric_layers.token_metric_layers import AudioFramesMetricLayer, SequenceTokenMetricLayer
 from neurst.metrics import build_metric
 from neurst.models import build_model
 from neurst.models.model_utils import deduce_text_length
@@ -378,10 +379,16 @@ class SpeechToText(Task):
                 window_size_func=window_size_fn))
 
     def build_metric_layer(self):
-        return [AudioFramesMetricLayer("src"), SequenceTokenMetricLayer("trg")]
+        return [AudioFramesMetricLayer("src"), SequenceTokenMetricLayer("trg"),
+                BatchCountMetricLayer("src")]
 
-    def get_eval_metric(self, args, name="metric"):
-        """ Returns a neurst.metrics.metric.Metric object for evaluation."""
+    def get_eval_metric(self, args, name="metric", ds=None):
+        """ Returns a bytedseq.metrics.metric.Metric object for evaluation."""
+        if ds is None:
+            return None
+        elif hasattr(ds, "trg_lang") and ds.trg_lang is not None:
+            return build_metric(args[name + ".class"], language=ds.trg_lang,
+                                **args[name + ".params"])
         return build_metric(args[name + ".class"], language=self._trg_data_pipeline.meta["language"],
                             **args[name + ".params"])